+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+- Added `cache` configuration to `GenerateConfig` (formerly was only available as a parameter to `generate()`).
+- ReAct agent: `on_continue` can now return a new `AgentState`.
+- OpenAI: Retries for `APIConnectionError`.
+
 ## 0.3.145 (09 November 2025)
 
 - OpenAI: Show reasoning summaries by default (auto-detect whether current account is capable of reasoning summaries and fallback as required).